Delete the source only when it was found

+++ tubesync/sync/tasks.py
@@ -792,8 +792,9 @@ def delete_all_media_for_source(source_id, source_name, source_directory):
         (source and source.delete_removed_media) or
         (directory_path / '.to_be_removed').is_file()
     )
-    with atomic(durable=True):
-        source.delete()
+    if source:
+        with atomic(durable=True):
+            source.delete()
     if remove:
         log.info(f'Deleting directory for: {source_name}: {directory_path}')
         rmtree(directory_path, True)
